# src/main/java/com/ntt/evaluation/user_manager/api/UserApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.ntt.evaluation.user_manager.api;

import com.ntt.evaluation.user_manager.model.Error;
import java.util.UUID;
import com.ntt.evaluation.user_manager.model.User;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-26T23:12:16.846028-03:00[America/Santiago]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "user", description = "Usuarios")
public interface UserApi {

    default UserApiDelegate getDelegate() {
        return new UserApiDelegate() {};
    }

    /**
     * POST /user : Crea un nuevo usuario
     * Crea un nuevo usuario
     *
     * @param user Crea un nuevo usuario (required)
     * @return Successful operation (status code 200)
     *         or Invalid input (status code 405)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "addUser",
        summary = "Crea un nuevo usuario",
        description = "Crea un nuevo usuario",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "405", description = "Invalid input"),
            @ApiResponse(responseCode = "default", description = "Unexpected error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/user",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<User> addUser(
        @Parameter(name = "User", description = "Crea un nuevo usuario", required = true) @Valid @RequestBody User user
    ) {
        return getDelegate().addUser(user);
    }


    /**
     * DELETE /user/{userId} : Borra un usuario existente
     * Borra un usuario existente
     *
     * @param userId Id del usuario a obtener (required)
     * @return Successful operation (status code 200)
     *         or Usuario no Encontrado (status code 404)
     *         or Invalid input (status code 405)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "deleteUser",
        summary = "Borra un usuario existente",
        description = "Borra un usuario existente",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation"),
            @ApiResponse(responseCode = "404", description = "Usuario no Encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "405", description = "Invalid input"),
            @ApiResponse(responseCode = "default", description = "Unexpected error")
        },
        security = {
            @SecurityRequirement(name = "usercrud_auth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/user/{userId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> deleteUser(
        @Parameter(name = "userId", description = "Id del usuario a obtener", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId
    ) {
        return getDelegate().deleteUser(userId);
    }


    /**
     * GET /user/{userId} : Obtiene un usuario por su id
     * Retorna. el usuario correspondiente al id
     *
     * @param userId Id del usuario a obtener (required)
     * @return El usuario (status code 200)
     *         or Usuario no Encontrado (status code 404)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "getUserById",
        summary = "Obtiene un usuario por su id",
        description = "Retorna. el usuario correspondiente al id",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "200", description = "El usuario", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "404", description = "Usuario no Encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error")
        },
        security = {
            @SecurityRequirement(name = "usercrud_auth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/user/{userId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<User> getUserById(
        @Parameter(name = "userId", description = "Id del usuario a obtener", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId
    ) {
        return getDelegate().getUserById(userId);
    }


    /**
     * GET /user : Lista los usarios del sistema
     * Lista los usarios del sistema
     *
     * @return Successful operation (status code 200)
     *         or No se encuentran usuarios (status code 404)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "listUsers",
        summary = "Lista los usarios del sistema",
        description = "Lista los usarios del sistema",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "404", description = "No se encuentran usuarios", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "default", description = "Unexpected error")
        },
        security = {
            @SecurityRequirement(name = "usercrud_auth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/user",
        produces = { "application/json" }
    )
    
    default ResponseEntity<User> listUsers(
        
    ) {
        return getDelegate().listUsers();
    }


    /**
     * PUT /user/{userId} : Actualiza un usuario existente
     * Actualiza un usuario existente
     *
     * @param userId Id del usuario a obtener (required)
     * @param user Actualiza un usuario existente (required)
     * @return Successful operation (status code 201)
     *         or Invalid input (status code 405)
     *         or Unexpected error (status code 200)
     */
    @Operation(
        operationId = "updateUser",
        summary = "Actualiza un usuario existente",
        description = "Actualiza un usuario existente",
        tags = { "user" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Successful operation"),
            @ApiResponse(responseCode = "405", description = "Invalid input"),
            @ApiResponse(responseCode = "default", description = "Unexpected error")
        },
        security = {
            @SecurityRequirement(name = "usercrud_auth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/user/{userId}",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> updateUser(
        @Parameter(name = "userId", description = "Id del usuario a obtener", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId,
        @Parameter(name = "User", description = "Actualiza un usuario existente", required = true) @Valid @RequestBody User user
    ) {
        return getDelegate().updateUser(userId, user);
    }

}